+++ src/components/camera/CenterHere.vue
@@ -99,7 +99,9 @@ const marker = ref({
 const newRa = ref(0);
 const newDec = ref(0);
 
-onMounted(() => {
+onMounted(async () => {
+  await fetchFramingInfo();
+
   window.addEventListener('resize', onWindowResize);
 
   // Beobachtet die Bildgröße
@@ -133,6 +135,14 @@ onMounted(() => {
 onBeforeUnmount(() => {
   window.removeEventListener('resize', onWindowResize);
 });
+async function fetchFramingInfo() {
+  try {
+    const data = await apiService.framingAction('info');
+    framingStore.framingInfo = data.Response;
+  } catch (error) {
+    console.error('Fehler beim Abrufen des FramingInfo:', error);
+  }
+}
 
 async function unparkMount() {
   if (store.mountInfo.AtPark) {
@@ -232,16 +242,30 @@ function keepTargetInBounds() {
  * calculateRaDec():
  * Ermittelt RA/Dec basierend auf Box-Position + Rotation.
  */
-function calculateRaDec() {
+async function calculateRaDec() {
   const rect = imageRef.value?.getBoundingClientRect();
   if (!rect) return;
 
-  const sensorWidth = store.cameraInfo.XSize;
+  let sensorWidth = store.cameraInfo.XSize;
+  let sensorHeight = store.cameraInfo.YSize;
+
+  sensorWidth = -1;
+
+  if (sensorWidth === -1) {
+    console.log('DLSR erkannt');
+    await fetchFramingInfo();
+    sensorWidth = framingStore.framingInfo.CameraWidth;
+    sensorHeight = framingStore.framingInfo.CameraHeight;
+  }
+
   const displayedWidth = rect.width;
   const ratioX = sensorWidth / displayedWidth;
-  const sensorHeight = store.cameraInfo.YSize;
   const displayedHeight = rect.height;
   const ratioY = sensorHeight / displayedHeight;
+
+  console.log(
+    `Chipgröße org in Pixeln: ${Math.round(store.cameraInfo.XSize)} x ${Math.round(store.cameraInfo.YSize)}`
+  );
 
   // Bildmitte
   const centerX = rect.width / 2;